Drop support for Python 3.5
- Switch to pathlib
- Stop using OrderedDict
- Switch to using f-strings

In `test_make_betaincinv22_table`, we were mistakenly comparing
the created test table to itself, rather than to the table to be
deployed. However, comparing against the deployment table can
result in errors, since different Numpy versions may result in
different table files. So we just remove this comparison, since
we already individually test the accuracy of the created table
and the deployment table.